+++ spark/src/main/spark-4.0/org/apache/comet/shims/ShimSQLConf.scala
@@ -24,13 +24,6 @@ import org.apache.spark.sql.internal.LegacyBehaviorPolicy
 
 trait ShimSQLConf {
 
-  /**
-   * Spark 3.4 renamed parquetFilterPushDownStringStartWith to
-   * parquetFilterPushDownStringPredicate
-   *
-   * TODO: delete after dropping Spark 3.2 & 3.3 support and simply use
-   * parquetFilterPushDownStringPredicate
-   */
   protected def getPushDownStringPredicate(sqlConf: SQLConf): Boolean =
     sqlConf.parquetFilterPushDownStringPredicate
