+++ spark/src/main/spark-4.0/org/apache/spark/sql/comet/ShimCometScanExec.scala
@@ -17,23 +17,22 @@
  * under the License.
  */
 
-package org.apache.comet.shims
-
-import scala.language.implicitConversions
+package org.apache.spark.sql.comet
 
 import org.apache.hadoop.fs.Path
 
-import org.apache.spark.{SparkContext, SparkException}
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.AttributeReference
 import org.apache.spark.sql.connector.read.{InputPartition, PartitionReaderFactory}
-import org.apache.spark.sql.execution.{FileSourceScanExec, PartitionedFileUtil}
-import org.apache.spark.sql.execution.datasources.{FilePartition, FileScanRDD, FileStatusWithMetadata, PartitionDirectory, PartitionedFile}
+import org.apache.spark.sql.errors.QueryExecutionErrors
 import org.apache.spark.sql.execution.datasources.parquet.ParquetOptions
 import org.apache.spark.sql.execution.datasources.v2.DataSourceRDD
+import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.execution.metric.SQLMetric
-import org.apache.spark.sql.types.{LongType, StructField, StructType}
+import org.apache.spark.sql.execution.{FileSourceScanExec, PartitionedFileUtil}
+import org.apache.spark.sql.types.StructType
+import org.apache.spark.SparkContext
 
 trait ShimCometScanExec {
   def wrapped: FileSourceScanExec
@@ -44,24 +43,16 @@ trait ShimCometScanExec {
     .map { a => a.setAccessible(true); a }
     .flatMap(_.invoke(wrapped).asInstanceOf[Seq[AttributeReference]])
 
-  // TODO: remove after dropping Spark 3.2 and 3.3 support and directly call
-  //       wrapped.fileConstantMetadataColumns
   lazy val fileConstantMetadataColumns: Seq[AttributeReference] =
-    wrapped.getClass.getDeclaredMethods
-      .filter(_.getName == "fileConstantMetadataColumns")
-      .map { a => a.setAccessible(true); a }
-      .flatMap(_.invoke(wrapped).asInstanceOf[Seq[AttributeReference]])
+    wrapped.fileConstantMetadataColumns
 
-  // TODO: remove after dropping Spark 3.2 support and directly call new DataSourceRDD
   protected def newDataSourceRDD(
       sc: SparkContext,
       inputPartitions: Seq[Seq[InputPartition]],
       partitionReaderFactory: PartitionReaderFactory,
       columnarReads: Boolean,
-      customMetrics: Map[String, SQLMetric]): DataSourceRDD = {
-    implicit def flattenSeq(p: Seq[Seq[InputPartition]]): Seq[InputPartition] = p.flatten
+      customMetrics: Map[String, SQLMetric]): DataSourceRDD =
     new DataSourceRDD(sc, inputPartitions, partitionReaderFactory, columnarReads, customMetrics)
-  }
 
   // TODO: remove after dropping Spark 3.2 support and directly call new FileScanRDD
   protected def newFileScanRDD(
@@ -89,40 +80,11 @@ trait ShimCometScanExec {
       .last
       .asInstanceOf[FileScanRDD]
 
-  // TODO: remove after dropping Spark 3.2 and 3.3 support and directly call
-  //       QueryExecutionErrors.SparkException
-  protected def invalidBucketFile(path: String, sparkVersion: String): Throwable = {
-    if (sparkVersion >= "3.3") {
-      val messageParameters = if (sparkVersion >= "3.4") Map("path" -> path) else Array(path)
-      classOf[SparkException].getDeclaredConstructors
-        .filter(_.getParameterCount == 3)
-        .map(_.newInstance("INVALID_BUCKET_FILE", messageParameters, null))
-        .last
-        .asInstanceOf[SparkException]
-    } else { // Spark 3.2
-      new IllegalStateException(s"Invalid bucket file ${path}")
-    }
-  }
+  protected def invalidBucketFile(path: String, sparkVersion: String): Throwable =
+    QueryExecutionErrors.invalidBucketFile(path)
 
-  // Copied from Spark 3.4 RowIndexUtil due to PARQUET-2161 (tracked in SPARK-39634)
-  // TODO: remove after PARQUET-2161 becomes available in Parquet
-  private def findRowIndexColumnIndexInSchema(sparkSchema: StructType): Int = {
-    sparkSchema.fields.zipWithIndex.find { case (field: StructField, _: Int) =>
-      field.name == ShimFileFormat.ROW_INDEX_TEMPORARY_COLUMN_NAME
-    } match {
-      case Some((field: StructField, idx: Int)) =>
-        if (field.dataType != LongType) {
-          throw new RuntimeException(
-            s"${ShimFileFormat.ROW_INDEX_TEMPORARY_COLUMN_NAME} must be of LongType")
-        }
-        idx
-      case _ => -1
-    }
-  }
-
-  protected def isNeededForSchema(sparkSchema: StructType): Boolean = {
-    findRowIndexColumnIndexInSchema(sparkSchema) >= 0
-  }
+  // see SPARK-39634
+  protected def isNeededForSchema(sparkSchema: StructType): Boolean = false
 
   protected def getPartitionedFile(f: FileStatusWithMetadata, p: PartitionDirectory): PartitionedFile =
     PartitionedFileUtil.getPartitionedFile(f, p.values, 0, f.getLen)
