+++ spark/src/test/scala/org/apache/spark/sql/benchmark/CometReadBenchmark.scala
@@ -25,11 +25,12 @@ import scala.collection.JavaConverters._
 import scala.util.Random
 
 import org.apache.spark.benchmark.Benchmark
+import org.apache.spark.comet.ShimTestUtils
 import org.apache.spark.sql.execution.datasources.parquet.VectorizedParquetRecordReader
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.vectorized.ColumnVector
 
-import org.apache.comet.{CometConf, TestUtils}
+import org.apache.comet.CometConf
 import org.apache.comet.parquet.BatchReader
 
 /**
@@ -123,7 +124,7 @@ object CometReadBenchmark extends CometBenchmarkBase {
             (col: ColumnVector, i: Int) => longSum += col.getUTF8String(i).toLongExact
         }
 
-        val files = TestUtils.listDirectory(new File(dir, "parquetV1"))
+        val files = ShimTestUtils.listDirectory(new File(dir, "parquetV1"))
 
         sqlBenchmark.addCase("ParquetReader Spark") { _ =>
           files.map(_.asInstanceOf[String]).foreach { p =>
